Made some small fixes and changes.

+++ Abilities/Gun.cs
@@ -7,8 +7,8 @@ namespace GameDevIdiotsProject1.Abilities
 {
 	class Gun : DefaultAbility
 	{
-		private const int COOLDOWN = 200;
-		private const int DURATION = 0;
+		private const float COOLDOWN = 200;
+		private const float DURATION = 0;
 		private const float BULLET_SPEED = .5f;
 
 		public Gun(Command command) : base(command)
